% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{alert_skip}
\alias{alert_skip}
\title{Alert if skip detected in given vector}
\usage{
alert_skip(x)
}
\arguments{
\item{x}{Numeric vector to be scanned}
}
\description{
Alert if skip detected in given vector
}
