% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ceasefire}
\alias{ceasefire}
\title{Make df with reference for early return}
\usage{
ceasefire(df, path, sheet, funcname)
}
\arguments{
\item{df}{Data frame to return}

\item{path}{File path to be used by \code{add_reference}}

\item{sheet}{Sheet name to be used by \code{add_reference}}

\item{funcname}{Name of function to create message for user}
}
\description{
Make df with reference for early return
}
