% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdate.R
\name{date2julian}
\alias{date2julian}
\title{Convert date to Excel Julian day}
\usage{
date2julian(x)
}
\arguments{
\item{x}{Date string in Gregorian or Japanes Calendar format.}
}
\value{
Julian day (Microsoft Excel style: origin = 1900-01-01).
}
\description{
Convert date to Excel Julian day
}
\examples{
date2julian("H30.01.01")
date2julian("H30-01-01")
date2julian("H30.1.1")
date2julian("H30-1-1")
}
