% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2rebel.R
\name{ensure_csv}
\alias{ensure_csv}
\title{Ensure the file name has specific extension}
\usage{
ensure_csv(fname)
}
\arguments{
\item{fname}{File name to be processed}
}
\description{
Ensure the file name has specific extension
}
