% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{extract_a_block}
\alias{extract_a_block}
\title{Extract a block from df using the keyword}
\usage{
extract_a_block(pos.key, find_from, direction, df, offset = c(0, 0), dim)
}
\arguments{
\item{pos.key}{Position where the \code{regex} of \code{extract_blocks()}
matched the keyword}

\item{find_from}{The row or column position
which \code{excract_block()} search key}

\item{direction}{The direction to which data blocks distribute}

\item{df}{Data frame to be processed}

\item{offset}{The offset (\code{c(row, pos})) of the block topleft from
the coordination of keyword}

\item{dim}{Dimension (\code{c(row, col)}) of the block}
}
\description{
This function is the substancial function of \code{extract_blocks()}.
}
