% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{extract_a_cluster}
\alias{extract_a_cluster}
\title{Extract a cluster from df using the keyword}
\usage{
extract_a_cluster(pos.key, find_from, direction, df, offset = c(0, 0),
  dim)
}
\arguments{
\item{pos.key}{Position where the \code{regex} of \code{extract_clusters()}
matched the keyword}

\item{find_from}{The row or column position
which \code{excract_cluster()} search key}

\item{direction}{The direction to which data clusters distribute}

\item{df}{Data frame to be processed}

\item{offset}{The offset (\code{c(row, pos})) of the cluster topleft from
the coordination of keyword}

\item{dim}{Dimension (\code{c(row, col)}) of the cluster}
}
\description{
This function is the substancial function of \code{extract_clusters()}.
}
