% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{extract_blocks}
\alias{extract_blocks}
\title{Extract data blocks from data frame using the keyword}
\usage{
extract_blocks(df, regex, col = NULL, row = NULL, offset = c(0, 0),
  dim)
}
\arguments{
\item{df}{Data frame to be processed}

\item{regex}{Regular expression to match keywords}

\item{col}{Column position from which the keyword to be searched}

\item{row}{Row position from which the keyword to be searched}

\item{offset}{The offset (\code{c(row, pos})) of the block topleft from
the coordination of keyword}

\item{dim}{Dimension (\code{c(row, col)}) of the block}
}
\description{
Extract data blocks from data frame using the keyword
}
