% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract data clusters from data frame using the keyword}
\usage{
extract_clusters(df, regex, col = NULL, row = NULL, offset = c(0, 0),
  dim, info = NULL)
}
\arguments{
\item{df}{Data frame to be processed}

\item{regex}{Regular expression to match keywords}

\item{col}{Column position from which the keyword to be searched}

\item{row}{Row position from which the keyword to be searched}

\item{offset}{The offset (\code{c(row, pos})) of the cluster topleft from
the coordination of keyword}

\item{dim}{Dimension (\code{c(row, col)}) of the cluster}

\item{info}{Parameters to control \code{link{append_info}}}
}
\description{
This function extracts data clusters from single Excel sheet.
}
