% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{fill_colhead}
\alias{fill_colhead}
\title{Fill NAs of merged column headers by left element}
\usage{
fill_colhead(df, rows = 1, regex = ".+")
}
\arguments{
\item{df}{Data frame to be processed}

\item{rows}{Row positions of the cells to be processed}

\item{regex}{Regex matches varname for filling}
}
\description{
Fill NAs of merged column headers by left element
}
