% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{fill_colname_topleft}
\alias{fill_colname_topleft}
\title{Coerce df topleft have namespace}
\usage{
fill_colname_topleft(df)
}
\arguments{
\item{df}{data frame to be processed}
}
\description{
Coerce df topleft have namespace
}
