% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmerge.R
\name{fill_na}
\alias{fill_na}
\title{Fill NAs of data frame by varname matched by regex}
\usage{
fill_na(df, along, pos)
}
\arguments{
\item{df}{Data frame to be processed}

\item{along}{\code{row} or \code{col}, along which NAs of df are filled up}

\item{pos}{Position of column or row to be processed}
}
\description{
Fill NAs of data frame by varname matched by regex
}
