% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{fill_rowhead}
\alias{fill_rowhead}
\title{Fill NAs of merged rows by 'varname'}
\usage{
fill_rowhead(df, cols = NULL, regex = ".+")
}
\arguments{
\item{df}{Data frame to be processed}

\item{cols}{Col positions of the cells to be filled by 'varname'}

\item{regex}{Regex matches varname for filling}
}
\description{
Fill NAs of merged rows by 'varname'
}
