% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{gather_cols}
\alias{gather_cols}
\title{Gather columns to variable}
\usage{
gather_cols(df, regex, newname, varname)
}
\arguments{
\item{df}{Data frame to be processed}

\item{regex}{Regex to match columns to be gathered}

\item{newname}{New name for colnames to be gatherd}

\item{varname}{New name for values to be gatherd}
}
\description{
Gather columns to variable
}
