% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdate.R
\name{get_datefmt}
\alias{get_datefmt}
\title{Get format of given date string}
\usage{
get_datefmt(x, year)
}
\arguments{
\item{x}{Date string to be judged}

\item{year}{\%Y of date string}
}
\value{
Format of date string, one of '\%Y\%m\%d', '\%m\%d', or 'XLjday'.
}
\description{
Get format of given date string
}
