% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{headerize}
\alias{headerize}
\title{Change specific row into df header}
\usage{
headerize(df, row)
}
\arguments{
\item{df}{Data frame to be processed}

\item{row}{Position of the row to make df header}
}
\description{
Change specific row into df header
}
