% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2rebel.R
\name{is_csv}
\alias{is_csv}
\title{Indicate whether the given fname has '.csv' extension}
\usage{
is_csv(fname)
}
\arguments{
\item{fname}{File name to be tested}
}
\description{
Indicate whether the given fname has '.csv' extension
}
