% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdate.R
\name{jpdate2julian}
\alias{jpdate2julian}
\title{Convert Japanese Calendar date to Excel Julian day}
\usage{
jpdate2julian(x)
}
\arguments{
\item{x}{Date string in Japanes Calendar format.}
}
\value{
Julian day (Microsoft Excel style: origin = 1900-01-01).
}
\description{
Convert Japanese Calendar date to Excel Julian day
}
\examples{
jpdate2julian("H30.01.01")
jpdate2julian("H30-01-01")
jpdate2julian("H30.1.1")
jpdate2julian("H30-1-1")
}
