% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{jpyr2ad}
\alias{jpyr2ad}
\title{Convert vectors composed only numeric jpyear without names of the eras.}
\usage{
jpyr2ad(x, start)
}
\arguments{
\item{x}{numeric jpyear vectors}

\item{start}{Name of the era}
}
\description{
Convert vectors composed only numeric jpyear without names of the eras.
}
