% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{list2df}
\alias{list2df}
\title{Expand single value list to data frame}
\usage{
list2df(list, nrow)
}
\arguments{
\item{list}{List with `key = value` pairs}

\item{nrow}{Nrows of df to be created}
}
\description{
Expand single value list to data frame
}
