% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2rebel.R
\name{load_param}
\alias{load_param}
\title{Load parameters to drive rebel() from .csv file}
\usage{
load_param(fname)
}
\arguments{
\item{fname}{file path to param sheet}
}
\description{
Load parameters to drive rebel() from .csv file
}
