% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{locate_matchend}
\alias{locate_matchend}
\title{Locate the end of repeated match}
\usage{
locate_matchend(str, regex)
}
\arguments{
\item{str}{String to be searched}

\item{regex}{Regex for search}
}
\description{
This function locates the end of the repeated matchs in string.
The first end of repeated match will be returned if there are
  multiple repeted match.
}
