% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{make_ascii}
\alias{make_ascii}
\title{Convert full-width numbers in df into ASCII numbers}
\usage{
make_ascii(df, col, numerize = FALSE)
}
\arguments{
\item{df}{Data frame to be processed}

\item{col}{Number of the target column}

\item{numerize}{If TRUE, remove characters convert column to numeric}
}
\description{
Convert full-width numbers in df into ASCII numbers
}
