% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{make_hougan}
\alias{make_hougan}
\title{Make vector \emph{houganshi}}
\usage{
make_hougan(str)
}
\arguments{
\item{str}{String vector with NA or multiple characters}
}
\value{
Long single string composed of single-word cell and whitespace
}
\description{
This function returns vector \emph{houganshi} to locate the position of
  target word (e.g. species name) in a Excel rows or columns.
To keep correspondence between nubmer of cells and nchar of output string,
This function replaces \code{NA} and a cell value with multiple characters.
}
\examples{
\dontrun{
  str <- rep(1:10, 10) \%>\%
    replace(which(. \%\% 3  == 0), NA) \%>\%
    replace(which(. \%\% 5  == 0), "foo") \%>\%
    as.character()
  make_hougan(str)
}
}
