% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{make_rect}
\alias{make_rect}
\title{Make loaded df as rectangular-shape df}
\usage{
make_rect(df, range)
}
\arguments{
\item{df}{Data frame to be processed}

\item{range}{Cell range to be extracted in Excel-format("A1:Z10")}
}
\description{
Make loaded df as rectangular-shape df
}
