% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{mcol2row}
\alias{mcol2row}
\title{Gather month columns to rows}
\usage{
mcol2row(df, varname)
}
\arguments{
\item{df}{Data frame with month column}

\item{varname}{Name of the new colname after gathering}
}
\description{
Gather month columns to rows
}
