% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2rebel.R
\name{mk_summary}
\alias{mk_summary}
\title{Make summary table of rebel result}
\usage{
mk_summary(df, params)
}
\arguments{
\item{df}{Output of \code{\link{rebel}}}

\item{params}{Parameter table parsed by \code{\link{load_param}}}
}
\description{
Make summary table of rebel result
}
