% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdate.R
\name{num2date}
\alias{num2date}
\title{Convert Julian date (Microsoft Excel style: origin = 1900-01-01)
  to \%Y-\%m-\%d}
\usage{
num2date(x)
}
\arguments{
\item{x}{Julian day read from Excel spredsheets.}
}
\value{
Date character in %Y-%m-%d format.
}
\description{
Convert Julian date (Microsoft Excel style: origin = 1900-01-01)
  to \%Y-\%m-\%d
}
\examples{
num2date(43466)
num2date(c(43466:43468, "43469"))
}
