% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{paste_rows}
\alias{paste_rows}
\title{Paste characters in multiple rows on the given columnn}
\usage{
paste_rows(col, rows, df)
}
\arguments{
\item{col}{Position of the target column}

\item{rows}{Rows to be concatenated}

\item{df}{Data frame to be processed}
}
\description{
Paste characters in multiple rows on the given columnn
}
