% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{pull_vector}
\alias{pull_vector}
\title{Pull vector out of data frame}
\usage{
pull_vector(df, direction, pos)
}
\arguments{
\item{df}{Data frame to be processed}

\item{direction}{Direction of the vector to be pulled from df
#' \describe{
 \item{"h"}{specific row of the df will be returned as avector}
 \item{"v"}{specific column of the df will be returned as a vector}
}}

\item{pos}{Row- or column position to be searched}
}
\description{
Pull vector out of data frame
}
