% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{rebel}
\alias{rebel}
\title{Rebel against godly Excel workbook}
\usage{
rebel(path, sheet_regex, row_merged = 0, col_merged = 0,
  cluster = NULL)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet_regex}{Regular expression to match sheetname}

\item{row_merged}{Row position of merged colnames}

\item{col_merged}{Column position of merged rownames}

\item{cluster}{List of parameters to control \code{\link{extract_clusters}}.
\describe{
  \item{dir}{direction of the cluster evolution}
  \item{pos}{row- or column- position of the key to locate the cluster}
  \item{regex}{same as that of \code{\link{extract_clusters}}}
  \item{offset}{same as that of \code{\link{extract_clusters}}}
  \item{dim}{same as that of \code{\link{extract_clusters}}}
}}
}
\description{
Rebel against godly Excel workbook
}
