% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{rep_na_rep}
\alias{rep_na_rep}
\title{Replace NAs in given vector by the repetition of the prior value}
\usage{
rep_na_rep(x)
}
\arguments{
\item{x}{Vector containing value and NA}
}
\description{
Replace NAs in given vector by the repetition of the prior value
}
