% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2rebel.R
\name{report}
\alias{report}
\title{Export result summary of rebel as csv file}
\usage{
report(df, params, reportf = NULL)
}
\arguments{
\item{df}{Output of \code{\link{rebel}}}

\item{params}{Parameter table parsed by \code{\link{load_param}}}

\item{reportf}{Name of report file to be exported}
}
\description{
Export result summary of rebel as csv file
}
