% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{rm_matchcol}
\alias{rm_matchcol}
\title{Remove columns matched to key}
\usage{
rm_matchcol(df, key, rowpos, regex)
}
\arguments{
\item{df}{Data frame to be processed}

\item{key}{Name of columns to be removed}

\item{rowpos}{Position of the row contains \code{key}}

\item{regex}{If TRUE, \code{key} was recognized as regular expression}
}
\description{
Remove columns matched to key
}
