% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{rm_matchrow}
\alias{rm_matchrow}
\title{Remove rows matched to key}
\usage{
rm_matchrow(df, key, colpos, regex)
}
\arguments{
\item{df}{Data frame to be processed}

\item{key}{Name of rows to be removed}

\item{colpos}{Position of the colmun contains key}

\item{regex}{If TRUE, \code{key} was recognized as regular expression}
}
\description{
Remove rows matched to key
}
