% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{rm_sumcol}
\alias{rm_sumcol}
\title{Remove summary cols from df}
\usage{
rm_sumcol(df, key, rowname, regex = FALSE)
}
\arguments{
\item{df}{Data frame to be processed}

\item{key}{Regex to detect summary cols}

\item{rowname}{Name of the row contains key}

\item{regex}{If TRUE, rowname was recognized by regular expression}
}
\description{
Remove summary cols from df
}
