% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{rm_sumrow}
\alias{rm_sumrow}
\title{Remove summary rows from df}
\usage{
rm_sumrow(df, key, colname)
}
\arguments{
\item{df}{Data frame to be processed}

\item{key}{Regex to detect summary rows}

\item{colname}{Name of the colmun contains key}
}
\description{
Remove summary rows from df
}
