% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{sheet2var}
\alias{sheet2var}
\title{Convert sheetname to variable}
\usage{
sheet2var(df, as)
}
\arguments{
\item{df}{Data frame to be processed}

\item{as}{Name of the new column which contains sheetnames}
}
\description{
Convert sheetname to variable
}
