% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdate.R
\name{split_jpdate}
\alias{split_jpdate}
\title{Split date str in japanese date format}
\usage{
split_jpdate(x)
}
\arguments{
\item{x}{Date string with year in Japanese era}
}
\description{
Split date str in japanese date format
}
\examples{
split_jpdate("H30.01.01")
split_jpdate("H.30.01.01")
}
