% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdate.R
\name{stdz_date}
\alias{stdz_date}
\title{Standardize date string in format "\%Y-\%m-\%d" format}
\usage{
stdz_date(x, year)
}
\arguments{
\item{x}{Date string to be judged}

\item{year}{\%Y of date string}
}
\value{
Date string standardized as "\%Y-\%m-\%d" format.
}
\description{
Standardize date string in format "\%Y-\%m-\%d" format
}
