% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2rebel.R
\name{tbl2rebel}
\alias{tbl2rebel}
\title{Drive rebel() loading parameters written in .csv file}
\usage{
tbl2rebel(objname, paramfname, reportf = NULL)
}
\arguments{
\item{objname}{Name of the Object choosen from 'object' column in param sheet}

\item{paramfname}{Path to the parameter sheet}

\item{reportf}{Name of the report file to export}
}
\description{
Drive rebel() loading parameters written in .csv file
}
