% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{unclusterize}
\alias{unclusterize}
\title{Extract data clusters from data frame using the keyword}
\usage{
unclusterize(df, regex, direction, pos, offset = c(0, 0), ends,
  info = NULL)
}
\arguments{
\item{df}{Data frame to be processed}

\item{regex}{Regular expression to match keywords}

\item{direction}{Directoin of the cluster revolution}

\item{pos}{Positon of row/column to scan using \code{regex}}

\item{offset}{The offset (\code{c(row, pos})) of the cluster topleft from
the coordination of keyword}

\item{ends}{List of regex to locate row- and column- ends of each cluster
Form should be like \code{ends = list(row = "2019", col = "[Dd]ecember$")}}

\item{info}{Parameters to control \code{link{append_info}}}
}
\description{
This function extracts data clusters from single Excel sheet.
}
