% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{unfiscal}
\alias{unfiscal}
\title{Convert fiscal year column into true year}
\usage{
unfiscal(df, ycol, mcol, month_start, rule)
}
\arguments{
\item{df}{Data frame to be processed}

\item{ycol}{Position of fiscal year column}

\item{mcol}{Position of month column}

\item{month_start}{Month where fiscal year starts}

\item{rule}{Rules to get true year.
Give \code{head} if true2019Oct is called fiscal2020Oct as in U.S.
Give \code{tail} if true2020Mar is called fiscal2019Mar as in Japan.}
}
\description{
Convert fiscal year column into true year
}
