% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{unfiscal_vec}
\alias{unfiscal_vec}
\title{Convert fiscal year to true year}
\usage{
unfiscal_vec(fisyr, month, month_start, rule)
}
\arguments{
\item{fisyr}{Fiscal year}

\item{month}{Month at \code{fisyr}}

\item{month_start}{Month where fiscal year starts}

\item{rule}{Rules to get true year.
Give \code{head} if true2019Oct is called fiscal2020Oct as in U.S.
Give \code{tail} if true2020Mar is called fiscal2019Mar as in Japan.}
}
\description{
Convert fiscal year to true year
}
