% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{unmerge_horiz}
\alias{unmerge_horiz}
\title{Fill NAs of merged columns by 'varname'}
\usage{
unmerge_horiz(df, row, regex = ".+")
}
\arguments{
\item{df}{Data frame to be processed}

\item{row}{Row position of the cells to be filled by 'varname'}

\item{regex}{Regex matches varname for filling}
}
\description{
Fill NAs of merged columns by 'varname'
}
