% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{unmerge_vert}
\alias{unmerge_vert}
\title{Fill NAs of merged rows by 'varname'}
\usage{
unmerge_vert(df, col, regex = ".+")
}
\arguments{
\item{df}{Data frame to be processed}

\item{col}{Col position of the cells to be filled by 'varname'}

\item{regex}{Regex matches varname for filling}
}
\description{
Fill NAs of merged rows by 'varname'
}
