% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{vectorize_row}
\alias{vectorize_row}
\title{Vectorize a row}
\usage{
vectorize_row(df, row)
}
\arguments{
\item{df}{Data frame to be processed}

\item{row}{Position of row to be vectorized}
}
\description{
Vectorize a row
}
