% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{which_decrease}
\alias{which_decrease}
\title{Return the location of decrese in given vector}
\usage{
which_decrease(x)
}
\arguments{
\item{x}{Numeric vector to be scanned}
}
\description{
Return the location of decrese in given vector
}
