% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{ycol2row}
\alias{ycol2row}
\title{Gather year column to rows}
\usage{
ycol2row(df, varname)
}
\arguments{
\item{df}{Data frame to be processed}

\item{varname}{New name for values, same as \code{\link[dplyr]{mutate}}}
}
\description{
Gather year column to rows
}
