% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_reference}
\alias{add_reference}
\title{Add columns 'fname' and 'sheet' to df}
\usage{
add_reference(df, fname, sheet)
}
\arguments{
\item{df}{Data frame to be processed}

\item{fname}{String to add df as column 'fname'}

\item{sheet}{String to add df as column 'sheet'}
}
\description{
Add columns 'fname' and 'sheet' to df
}
