% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ceasefire}
\alias{ceasefire}
\title{Make df with reference for early return}
\usage{
ceasefire(df, path = NULL, sheet = NULL, funcname, posnames = TRUE)
}
\arguments{
\item{df}{Data frame to return}

\item{path}{File path to be used by \code{add_reference}}

\item{sheet}{Sheet name to be used by \code{add_reference}}

\item{funcname}{Name of function to create message for user}

\item{posnames}{If TRUE, row- and col- names are set to 'pos=' format for
following cluster extraction}
}
\description{
Make df with reference for early return
}
