% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclusterize.R
\name{extract_a_cluster}
\alias{extract_a_cluster}
\title{Extract a cluster from df using the keyword}
\usage{
extract_a_cluster(pos_key, find_from, direction, df, offset = c(0, 0),
  ends, info = NULL)
}
\arguments{
\item{pos_key}{Position where the \code{regex} of \code{unclusterize}
matched the keyword}

\item{find_from}{The row or column position
which \code{excract_cluster()} search key}

\item{direction}{The direction to which data clusters distribute}

\item{df}{Data frame to be processed}

\item{offset}{The offset (\code{c(row, pos})) of the cluster topleft from
the coordination of keyword}

\item{ends}{List of regex to locate row- and column- ends of each cluster
  Form should be like \code{ends = list(row = "2019", col = "[Dd]ecember$")}.
Regex \code{row = } must specify the end of 'left most' columnn of df,
 not that of the column with key matched by \code{regex}}

\item{info}{Parameters to make key:value list such as
\describe{
 \item{key_offset}{Offset \code{c(row, col)} of \code{key} topleft
  from df topleft. If \code{NULL}, automatically set to \code{keyn}}
 \item{key_dim}{Dimension \code{c(row, col)} of \code{key}}
 \item{value_offset}{Offset \code{c(row, col)} of \code{value} topleft from
  df topleft}
 \item{value_dim}{Dimension \code{c(row, col)} of \code{value}}
}}
}
\description{
This function is the substancial function of \code{unclusterize}.
}
