% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{fill_rowhead}
\alias{fill_rowhead}
\title{Fill NAs of merged rows by 'varname'}
\usage{
fill_rowhead(df, cols = 1, regex = ".+")
}
\arguments{
\item{df}{Data frame to be processed}

\item{regex}{Regex matches varname for filling}

\item{col}{Col position of the cells to be filled by 'varname'}
}
\description{
Fill NAs of merged rows by 'varname'
}
