% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{locate_keys}
\alias{locate_keys}
\title{Locate keywords in row or column of the given data frame}
\usage{
locate_keys(df, row = NULL, col = NULL, regex, stop_at = NULL)
}
\arguments{
\item{df}{Data frame to be processed}

\item{row}{Row position of df where the keyword appears}

\item{col}{Column position of df where the keyword appears}

\item{regex}{Regex to match keyword}

\item{stop_at}{\code{regex} search stops if \{stop_at\} matched}
}
\description{
Locate keywords in row or column of the given data frame
}
