% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{make_ascii}
\alias{make_ascii}
\title{Convert full-width numbers in df into ASCII numbers}
\usage{
make_ascii(x, col = NULL, row = NULL, numerize = FALSE, headerized = FALSE)
}
\arguments{
\item{x}{Data frame or vector to be processed}

\item{col}{Number of the target column}

\item{row}{Number of the target row}

\item{numerize}{If TRUE, remove characters convert column to numeric}

\item{headerized}{If FALSE (default), allow df with tentative colnames}
}
\description{
Convert full-width numbers in df into ASCII numbers
}
