% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmerge.R
\name{merge_colname}
\alias{merge_colname}
\title{Merge colnames of multiple rows}
\usage{
merge_colname(df, rows, cols = NULL)
}
\arguments{
\item{df}{Data frame to be processed}

\item{rows}{Rows of the target colnames to be concatenated}

\item{cols}{Numbers of target columns if given}
}
\description{
Merge colnames of multiple rows
}
