% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{rebel}
\alias{rebel}
\title{Rebel against godly Excel workbook}
\usage{
rebel(path, sheet.regex, row_merged = 0, col_merged = 0,
  cluster = NULL)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet.regex}{Regular expression to match sheetname}

\item{row_merged}{Row position of merged colnames}

\item{col_merged}{Column position of merged rownames}

\item{cluster}{List of cluster, direction, pos}
}
\description{
Rebel against godly Excel workbook
}
