% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{rebel}
\alias{rebel}
\title{Rebel against godly Excel workbook}
\usage{
rebel(path, sheet_regex, row_merged = 0, col_merged = 0,
  cluster = NULL, row_type = NULL, col_type = NULL,
  row_omit = NULL, col_omit = NULL, fullwidth = NULL)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet_regex}{Regular expression to match sheetname}

\item{row_merged}{Row position of merged colnames}

\item{col_merged}{Column position of merged rownames}

\item{cluster}{List of parameters to control \code{\link{extract_clusters}}.
\describe{
  \item{dir}{direction of the cluster evolution either of
  \code{"h"} (horizontal) or \code{"v"} (vertical)}
  \item{pos}{row- or column- position of the key to locate the cluster}
  \item{regex}{same as that of \code{\link{extract_clusters}}}
  \item{offset}{same as that of \code{\link{extract_clusters}}}
  \item{dim}{same as that of \code{\link{extract_clusters}}}
  \item{info}{same as that of \code{\link{extract_clusters}}}
}}

\item{row_type}{Type of row one of
\describe{
 \item{Y}{year}
 \item{YM}{year-month}
 \item{jY}{Japanese year}
 \item{jYM}{Japanese year-month}
 \item{fisY}{fiscal year}
 \item{fisYM}{fiscal year-month}
 \item{fisjYM}{fiscal Japanese year-month}
}}

\item{col_type}{List of parameters to control \code{\link{gather_cols}}.}

\item{row_omit}{List of parameters to control \code{\link{rm_matchrow}}}

\item{col_omit}{List of parameters to control \code{\link{rm_matchcol}}}

\item{fullwidth}{List of parameters to cotrol \code{\link{make_ascii}}}
}
\description{
Rebel against godly Excel workbook
}
