% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{rebel_sheet}
\alias{rebel_sheet}
\title{Rebel against godly Excel worksheet}
\usage{
rebel_sheet(sheet, path, row_merged, col_merged)
}
\arguments{
\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{path}{Path to the xls/xlsx file.}

\item{row_merged}{Row position of merged colnames}

\item{col_merged}{Column position of merged rownames}
}
\description{
Rebel against godly Excel worksheet
}
