% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{rebel_sheet}
\alias{rebel_sheet}
\title{Rebel against godly Excel worksheet}
\usage{
rebel_sheet(sheet, path, row_merged = 0, col_merged = 0,
  cluster = NULL, row_type = NULL, col_type = NULL,
  row_omit = NULL, col_omit = NULL)
}
\arguments{
\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{path}{Path to the xls/xlsx file.}

\item{row_merged}{Row position of merged colnames}

\item{col_merged}{Column position of merged rownames}

\item{cluster}{List of parameters to control \code{\link{extract_clusters}}.
\describe{
  \item{dir}{direction of the cluster evolution}
  \item{pos}{row- or column- position of the key to locate the cluster}
  \item{regex}{same as that of \code{\link{extract_clusters}}}
  \item{offset}{same as that of \code{\link{extract_clusters}}}
  \item{dim}{same as that of \code{\link{extract_clusters}}}
}}

\item{row_type}{Type of row one of
\describe{
 \item{Y}{year}
 \item{YM}{year-month}
 \item{jY}{Japanese year}
 \item{jYM}{Japanese year-month}
 \item{fisY}{fiscal year}
 \item{fisYM}{fiscal year-month}
 \item{fisjYM}{fiscal Japanese year-month}
}}

\item{col_type}{list of parameters to control \code{\link{mcol2row}}.
\describe{
 \item{varname}{new varname, same as \code{value} of
                 \code{\link[tidyr]{gather}}}
}}

\item{row_omit}{list of parameters to control \code{\link{rm_sumrow}}}

\item{col_omit}{list of parameters to control \code{\link{rm_sumcol}}}
}
\description{
Rebel against godly Excel worksheet
}
