% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebel.R
\name{rebel_sheet}
\alias{rebel_sheet}
\title{Rebel against godly Excel worksheet}
\usage{
rebel_sheet(sheet, path, row_headers = NULL, col_headers = NULL,
  cluster = NULL, row_type = NULL, col_type = NULL,
  row_omit = NULL, col_omit = NULL, unfiscalize = c(month_start =
  NULL, rule = NULL), print_posnames = FALSE, crop = NULL)
}
\arguments{
\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{path}{Path to the xls/xlsx file.}

\item{row_headers}{Row position of merged colnames}

\item{col_headers}{Column position of merged rownames}

\item{cluster}{List of parameters to control \code{\link{unclusterize}}.
\describe{
  \item{dir}{direction of the cluster evolution either of
  \code{"h"} (horizontal) or \code{"v"} (vertical)}
  \item{pos}{row- or column- position of the key to locate the cluster}
  \item{regex}{same as that of \code{\link{unclusterize}}}
  \item{offset}{same as that of \code{\link{unclusterize}}}
  \item{dim}{same as that of \code{\link{unclusterize}}}
  \item{info}{same as that of \code{\link{unclusterize}}}
}}

\item{row_type}{Type of row one of
\describe{
 \item{Y}{year}
 \item{YM}{year-month}
 \item{jY}{Japanese year}
 \item{jYM}{Japanese year-month}
 \item{fisY}{fiscal year}
 \item{fisYM}{fiscal year-month}
 \item{fisjYM}{fiscal Japanese year-month}
}}

\item{col_type}{List of parameters to control \code{\link{gather_cols}}}

\item{row_omit}{List of parameters to control \code{\link{rm_matchrow}}}

\item{col_omit}{List of parameters to control \code{\link{rm_matchcol}}}

\item{unfiscalize}{List of parameters to control \code{\link{unfiscalize}}}

\item{print_posnames}{Same as that of \code{\link{ceasefire}}}

\item{crop}{Params from \code{crop()} in list format}
}
\description{
Rebel against godly Excel worksheet
}
