% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetools.R
\name{unclusterize}
\alias{unclusterize}
\title{Extract data clusters from data frame using the keyword}
\usage{
unclusterize(df, regex, direction, pos, offset = c(0, 0), ends,
  info = NULL, stop_at = NULL)
}
\arguments{
\item{df}{Data frame to be processed}

\item{regex}{Regular expression to match keywords}

\item{direction}{Directoin of the cluster revolution}

\item{pos}{Positon of row/column to scan using \code{regex}}

\item{offset}{The offset (\code{c(row, pos})) of the cluster topleft from
the coordination of keyword}

\item{ends}{List of regex to locate row- and column- ends of each cluster
  Form should be like \code{ends = list(row = "2019", col = "[Dd]ecember$")}.
Regex \code{row = } must specify the end of 'left most' columnn of df,
 not that of the column with key matched by \code{regex}}

\item{info}{Parameters to make key:value list such as
\describe{
 \item{key_offset}{Offset \code{c(row, col)} of \code{key} topleft
  from df topleft. If \code{NULL}, automatically set to \code{keyn}}
 \item{key_dim}{Dimension \code{c(row, col)} of \code{key}}
 \item{value_offset}{Offset \code{c(row, col)} of \code{value} topleft from
  df topleft}
 \item{value_dim}{Dimension \code{c(row, col)} of \code{value}}
}}

\item{stop_at}{Same as that of \code{\link{locate_keys}}}
}
\description{
This function extracts data clusters from single Excel sheet.
}
